+++ Argus/Source/Argus/ECS/Systems/TransformSystems.cpp
@@ -5,7 +5,7 @@
 
 void TransformSystems::RunSystems(float deltaTime)
 {
-	for (uint32 i = 0; i < ArgusECSConstants::k_maxEntities; ++i)
+	for (uint16 i = 0; i < ArgusECSConstants::k_maxEntities; ++i)
 	{
 		std::optional<ArgusEntity> potentialEntity = ArgusEntity::RetrieveEntity(i);
 		if (!potentialEntity.has_value())
